SDK-1463: Move filter to ID document
