+++ src/DocScan/Session/Create/Filters/RequiredDocument.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Yoti\DocScan\Session\Create\Filters;
 
-use Yoti\Util\Json;
-
 abstract class RequiredDocument implements \JsonSerializable
 {
     /**
@@ -14,18 +12,11 @@ abstract class RequiredDocument implements \JsonSerializable
     private $type;
 
     /**
-     * @var RequiredDocumentFilter|null
-     */
-    private $filter;
-
-    /**
      * @param string $type
-     * @param RequiredDocumentFilter|null $filter
      */
-    public function __construct(string $type, ?RequiredDocumentFilter $filter)
+    public function __construct(string $type)
     {
         $this->type = $type;
-        $this->filter = $filter;
     }
 
     /**
@@ -33,9 +24,8 @@ abstract class RequiredDocument implements \JsonSerializable
      */
     public function jsonSerialize(): \stdClass
     {
-        return (object) Json::withoutNullValues([
+        return (object) [
             'type' => $this->type,
-            'filter' => $this->filter,
-        ]);
+        ];
     }
 }
