+++ src/DocScan/Session/Create/Filters/RequiredIdentityDocumentBuilder.php
@@ -4,12 +4,28 @@ declare(strict_types=1);
 
 namespace Yoti\DocScan\Session\Create\Filters;
 
-class RequiredIdentityDocumentBuilder extends RequiredDocumentBuilder
+class RequiredIdentityDocumentBuilder
 {
     /**
-     * @inheritDoc
+     * @var RequiredDocumentFilter
      */
-    public function build(): RequiredDocument
+    protected $filter;
+
+    /**
+     * @param RequiredDocumentFilter $filter
+     *
+     * @return $this
+     */
+    public function withFilter(RequiredDocumentFilter $filter): self
+    {
+        $this->filter = $filter;
+        return $this;
+    }
+
+    /**
+     * @return RequiredIdentityDocument
+     */
+    public function build(): RequiredIdentityDocument
     {
         return new RequiredIdentityDocument($this->filter);
     }
