+++ tests/DocScan/Session/Create/Filters/RequiredDocumentTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Yoti\Test\DocScan\Session\Create\Filters;
 
 use Yoti\DocScan\Session\Create\Filters\RequiredDocument;
-use Yoti\DocScan\Session\Create\Filters\RequiredDocumentFilter;
 use Yoti\Test\TestCase;
 
 /**
@@ -21,22 +20,16 @@ class RequiredDocumentTest extends TestCase
      * @covers ::__construct
      * @covers ::jsonSerialize
      */
-    public function shouldSerializeWithTypeAndFilter()
+    public function shouldSerializeWithType()
     {
-        $requiredDocumentFilterMock = $this->createMock(RequiredDocumentFilter::class);
-        $requiredDocumentFilterMock
-            ->method('jsonSerialize')
-            ->willReturn((object) ['some' => 'filter']);
-
         $requiredDocument = $this->getMockBuilder(RequiredDocument::class)
-            ->setConstructorArgs([self::SOME_TYPE, $requiredDocumentFilterMock])
+            ->setConstructorArgs([self::SOME_TYPE])
             ->setMethodsExcept(['jsonSerialize'])
             ->getMock();
 
         $this->assertJsonStringEqualsJsonString(
             json_encode((object) [
                 'type' => self::SOME_TYPE,
-                'filter' => $requiredDocumentFilterMock,
             ]),
             json_encode($requiredDocument)
         );
